move custom defichain evm chains into separate file
